# [proto] Small improvement for tensor equalize op