@vfdev-5 2022-10-11
There can be more improvement if we vectorize histogram computation with `scatter_add_` (cc @lezcano). 
As Vasilis is suggesting to go iteratively, let's merge this PR first and put other improvements in follow-up PRs.